+++ module/actor/actor-covenant-sheet.js
@@ -317,7 +317,7 @@ export class ArM5eCovenantActorSheet extends ArM5eActorSheet {
     if (type == "spell" || type == "magicalEffect" || type == "enchantment") {
       log(false, "Valid drop");
       // create a labText data:
-      return await super._onDropItemCreate(effectToLabText(foundry.utils.deepClone(item)));
+      return await super._onDropItemCreate(effectToLabText(item.toObject()));
     }
     // }
     const res = await super._onDropItem(event, data);